Split automation scheduler orchestration
